Rotation CV ++
Added a filter and some UIbuttons for Region Filtering to the main controller. Should add the display view. Also, I need to handle an empty datasource when the layout is being generated, it seems the layout is trying to apply attributes even if there isn't an object, I suspect this has to do with the start/end index of the attribute map function.